scripts/requirements.txt
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
scipy>=1.10.0
noisereduce>=3.0.0
numpy>=1.24.0
